Add extra bias away from obstacles
This commit adds a stong bias against being closer than a certain threshold
to obstacles, and gives a reward for being farther than a certain threshold
from obstacles. This reduces the tendency to travel too close to walls and
corners.

Some new documentation and command-line options were added as well, to
improve configurability.