Improve fast-computing mode CSV output
This commit adds new parameters to the CSV output of the --fast-computing
mode, and changes the CSV separator to commas (,) instead of pipes (|).

Also:

- The main game loop has also been refactored into two loops, one for fast
computing mode and the other for normal mode, to improve readbility.

- The method of storing robot statistics such as the number of
glitches has been changed to use a separate class, to make it easier to
access statistical information.

- Map 'a8.jpg' has been converted to 'a8.png' to fix the bug where JPEG
  compression caused obstacles not to be detected properly.