Fix obstacle-tunneling glitch
This commit attempts to stop the robot from tunneling through
obstacles. From testing, it appears that the robot no longer tunnels
through obstacles at all, except for when a dynamic obstacle goes
through the robot. Even in that case, the robot usually returns to a
free space when the dynamic obstacle moves past.