Add more configurability for memory and upgrade speed modes
There are several changes in this commit:

1. Options were added to improve memory control. They are
`--robot-memory-sigma`, `--robot-memory-decay`, and `--robot-memory-size`.
Sigma controls the memory strength of a point with respect to distance,
decay does the same but with respect to time, and size is the maximum
number of memory points that will be considered.

2. The display updates were removed, since running cProfile on different
machines showed that while rendering isn't a big factor on some, it can
cause a noticable performance decrease on others. Display updates are
therefore disabled by default when running in `--batch-mode`.

3. New key commands were added. Now, pressing 'e' while a simulation is
running will toggle whether the display updates every frame (useful when
running in `--batch-mode`, where display updates are turned off by
default). Pressing 'u' will update the frame just once. Pressing 'q'
will quit the simulation.

4. Safe speed settings were tuned. See the diff for full details, but
the most significant changes are that movement direction and memory
points are no longer affected by the dynamic obstacle check. Only the
speed will change according to dynamic obstacles, and in essence, the
safe robot will no longer slow down at all, only speed up when an
obstacle is behind it.