Rewrite dynamic obstacle scanner
This commit rewrites the dynamic obstacle scanner for the robot radar. The
scans now use a direct equation to check whether each radar beam intersects
with each object, as opposed to the previous method which was
sampling-based, and checked to see if certain points along each beam were
pixel-colored.

The new method makes it possible to avoid having to redraw dynamic obstacles
when the display was turned off, and improves accuracy by obtaining the
exact point of intersection rather than using guess-and-check.