Make navigation algorithm more abstract and modular
This commit adds an abstract navigation algorithm superclass, to
better define the interface. It moves the concrete implementation
to a `FuzzyNavigationAlgorithm` class to make it possible to swap
in other navigation algorithm implementations that use the same
interface.

The changes to the `doxygen.conf` file are to add the setting to
traverse subdirectories, so the new package can be properly
documented.