Reduce strong coupling between robots and navigation algorithms
Each navigation algorithm previously took an instance of Robot in its
constructor, and was therefore very dependent on the particular field
names in robot. In addition, a Robot would create its own navigation
algorithm rather than allowing the constructor of the Robot to specify
its algorithm.

Now, each Robot has sensors, which are assigned by the constructor of
the robot. A Robot does not create its own navigation algorithm, but
instead the constructor of the Robot will create the navigation
algorithm. Each NavigationAlgorithm takes sensors and a target point
in the constructor, so there is no longer any dependency on Robot.