Move Game setup code to new file
There was a lot of configuration code in `Game.__init__` that had to
get overwritten every time the simulation setup changed, making it
hard to manage configurations when switching between different types
of simulations. This commit moves most of that code to a new file
`run_default.py`, leaving `Game.py` with the code that is common to
almost all simulations.

Some command-line options were removed as a result of the refactor.